# Conflicts:
#	Cargo.lock
#	Cargo.toml
#	crates/vox-orchestrator/src/jj_backend.rs
#	crates/vox-vcs/Cargo.toml
#	crates/vox-vcs/src/backend.rs
#	crates/vox-vcs/src/cas_fallback.rs
#	crates/vox-vcs/src/lib.rs
#	docs/src/architecture/layers.toml
#	docs/superpowers/plans/2026-06-05-jj-first-class-p2-jj-engine-and-dead-code-removal.md

+++ crates/vox-db/src/store/ops_orchestrator.rs
@@ -144,7 +144,7 @@ impl crate::VoxDb {
                         (),
                     )
                     .await?;
-                Ok(rows_affected as u64)
+                Ok(rows_affected)
             })
             .await
     }
@@ -236,7 +236,7 @@ impl crate::VoxDb {
                         params![min_seen_ms],
                     )
                     .await?;
-                Ok(affected as u64)
+                Ok(affected)
             })
             .await
     }
